fix: Remove conditional rendering logic from components
BREAKING CHANGE: The components in the library are not responsible anymore for the conditional rendering derived from the window size. Instead it is suggested to pass a prop on the client side to determine when the component should show a mobile version or when its grid has to be fluid.